+++ src/main/java/io/oeid/mogakgo/domain/project_join_req/domain/entity/ProjectJoinRequest.java
@@ -91,4 +91,5 @@ public class ProjectJoinRequest {
         if (!sender.getId().equals(user.getId())) {
             throw new ProjectJoinRequestException(PROJECT_JOIN_REQUEST_FORBIDDEN_OPERATION);
         }
+    }
 }
