+++ src/main/java/io/oeid/mogakgo/domain/project_join_req/infrastructure/ProjectJoinRequestRepositoryCustom.java
@@ -3,7 +3,7 @@ package io.oeid.mogakgo.domain.project_join_req.infrastructure;
 import io.oeid.mogakgo.common.base.CursorPaginationInfoReq;
 import io.oeid.mogakgo.common.base.CursorPaginationResult;
 import io.oeid.mogakgo.domain.project_join_req.domain.entity.enums.RequestStatus;
-import io.oeid.mogakgo.domain.project_join_req.presentation.projectJoinRequestRes;
+import io.oeid.mogakgo.domain.project_join_req.presentation.dto.res.projectJoinRequestRes;
 
 public interface ProjectJoinRequestRepositoryCustom {
